Test de envio de estructuras con sockets
